+++ packages/orq-rc/src/models/components/partdoneevent.ts
@@ -79,7 +79,7 @@ export const PartReasoningPart$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  _id: z.string().default("reasoning_01kc8y899wb3qzs0mtdyrbjn63"),
+  _id: z.string().default("reasoning_01kc9bp76x9vqx0j7s1x67m62t"),
   metadata: z.record(z.any()).optional(),
   kind: PartKind$inboundSchema,
   reasoning: z.string(),
@@ -104,7 +104,7 @@ export const PartReasoningPart$outboundSchema: z.ZodType<
   z.ZodTypeDef,
   PartReasoningPart
 > = z.object({
-  id: z.string().default("reasoning_01kc8y899wb3qzs0mtdyrbjn63"),
+  id: z.string().default("reasoning_01kc9bp76x9vqx0j7s1x67m62t"),
   metadata: z.record(z.any()).optional(),
   kind: PartKind$outboundSchema,
   reasoning: z.string(),