+++ packages/orq-rc/src/models/operations/createconversation.ts
@@ -272,7 +272,7 @@ export const CreateConversationResponseBody$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  _id: z.string().default("conv_01kc8y89hrxwrw5h0c13vy02nr"),
+  _id: z.string().default("conv_01kc9bp7f1q1dkaznymh79504e"),
   entity_id: z.string(),
   kind: CreateConversationKind$inboundSchema,
   display_name: z.string(),
@@ -312,7 +312,7 @@ export const CreateConversationResponseBody$outboundSchema: z.ZodType<
   z.ZodTypeDef,
   CreateConversationResponseBody
 > = z.object({
-  id: z.string().default("conv_01kc8y89hrxwrw5h0c13vy02nr"),
+  id: z.string().default("conv_01kc9bp7f1q1dkaznymh79504e"),
   entityId: z.string(),
   kind: CreateConversationKind$outboundSchema,
   displayName: z.string(),