+++ packages/orq-rc/src/models/operations/streamrunagent.ts
@@ -4091,7 +4091,7 @@ export const AgentToolInputRunTools$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  id: z.string().default("01KC8Y89FWNY82P5TAEYJ320V5"),
+  id: z.string().default("01KC9BP7DGJKHVBSVZWT2P6K7B"),
   name: z.string(),
   description: z.string().optional(),
   schema: z.lazy(() => AgentToolInputRunSchema$inboundSchema),
@@ -4110,7 +4110,7 @@ export const AgentToolInputRunTools$outboundSchema: z.ZodType<
   z.ZodTypeDef,
   AgentToolInputRunTools
 > = z.object({
-  id: z.string().default("01KC8Y89FWNY82P5TAEYJ320V5"),
+  id: z.string().default("01KC9BP7DGJKHVBSVZWT2P6K7B"),
   name: z.string(),
   description: z.string().optional(),
   schema: z.lazy(() => AgentToolInputRunSchema$outboundSchema),