+++ packages/orq-rc/RELEASES.md
@@ -5559,3 +5559,13 @@ Based on:
 - [typescript v4.1.0-rc.40] packages/orq-rc
 ### Releases
 - [NPM v4.1.0-rc.40] https://www.npmjs.com/package/@orq-ai/node/v/4.1.0-rc.40 - packages/orq-rc
+
+## 2025-12-12 13:25:18
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.675.4 (2.779.2) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.1.0-rc.41] packages/orq-rc
+### Releases
+- [NPM v4.1.0-rc.41] https://www.npmjs.com/package/@orq-ai/node/v/4.1.0-rc.41 - packages/orq-rc